moved custom errors in riot package into own package
