+++ riot/errors/errors_test.go
@@ -1,4 +1,4 @@
-package riot
+package errors
 
 import (
 	"bytes"
